Mudando nome da connection do firebase
